Moved Firebase out of components folder for clarity. Change AuthCondition to authCondition to adhere to standards
