+++ src/components/Session/index.js
@@ -1,5 +1,5 @@
 import AuthUserContext from './context';
 import withAuthentication from './withAuthentication';
-import withAuthorization, { AuthCondition } from './withAuthorization';
+import withAuthorization, { authCondition } from './withAuthorization';
 
-export { AuthUserContext, withAuthentication, withAuthorization, AuthCondition };
+export { AuthUserContext, withAuthentication, withAuthorization, authCondition };
