+++ src/components/Session/withAuthorization.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withRouter } from 'react-router-dom';
 
 import AuthUserContext from './context';
-import { withFirebase } from '../Firebase';
+import { withFirebase } from '../../Firebase';
 import * as ROUTES from '../../constants/routes';
 
 const withAuthorization = condition => Component => {
@@ -34,8 +34,8 @@ const withAuthorization = condition => Component => {
   return withRouter(withFirebase(WithAuthorization));
 };
 
-const AuthCondition = authUser => !!authUser;
+const authCondition = authUser => !!authUser;
 
 export default withAuthorization;
 
-export { AuthCondition };
+export { authCondition };
