+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
-import Firebase, { FirebaseContext } from './components/Firebase/';
+import Firebase, { FirebaseContext } from './Firebase/';
 
 ReactDOM.render(
     <FirebaseContext.Provider value={new Firebase()}>
